% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_game.R
\name{analyze_game}
\alias{analyze_game}
\title{Analyze a chess game}
\usage{
analyze_game(engine, san = NULL, lan = NULL, quiet = FALSE, ...)
}
\arguments{
\item{engine}{Either a character string giving the path to a chess engine, or
an engine handler created by \code{\link[=uci_engine]{uci_engine()}}.}

\item{san}{(default = \code{NULL}) A character string giving a series of chess
moves in short algebraic notation.}

\item{lan}{(default = \code{NULL}) A character string giving a series of chess
moves in UCI long algebraic notation.}

\item{quiet}{(default = \code{FALSE}) A Boolean. Hide system messages?}

\item{...}{Further arguments passed to \code{\link[=uci_go]{uci_go()}}.}
}
\value{
A list containing the results from \code{analyze_position()} for each
position in the game.
}
\description{
This function analyzes a chess game using a
\href{https://wbec-ridderkerk.nl/html/UCIProtocol.html}{UCI-compatible} chess
engine.
}
\details{
The moves of the game may be provided in either short algebraic
notation (SAN) or UCI long algebraic notation (LAN). Note that if it is
black's turn, the score is multiplied by -1 so that a positive score always
means white has the advantage, and negative always means black has the
advantage.
}
\examples{
# To run the example code, place a UCI-compatible chess engine in the
# `bigchess` subdirectory `/inst/extdata/engine`, or replace the line below
# with a valid path:
# engine_path <- "/put/your/own/engine/path/here"
engine_path <- find_engine()
if(!is.null(engine_path)) {
  san <- "1. e4 e5 2. Nf3 Nc6 3. d4 exd4 4. Bc4 Nf6 5. O-O Be7"
  game <- analyze_game(engine_path, san = san, depth = 2)
  game[[1]] # Analysis results for the first position
} else {
 message(
   paste0(
     'To run the examples, install a chess engine in /inst/extdata/',
     'engine,\n or replace engine_path with the path to an engine.'
   )
 )
}
}
\seealso{
\code{\link[=analyze_position]{analyze_position()}}
}
