% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_position.R
\name{analyze_position}
\alias{analyze_position}
\title{Analyze a chess position}
\usage{
analyze_position(engine, san = NULL, lan = NULL, ...)
}
\arguments{
\item{engine}{Either a character string giving the path to a chess engine, or
an engine handler created by \code{\link[=uci_engine]{uci_engine()}}.}

\item{san}{(default = \code{NULL}) A character string giving a series of chess
moves in short algebraic notation.}

\item{lan}{(default = \code{NULL}) A character string giving a series of chess
moves in UCI long algebraic notation.}

\item{...}{Further arguments passed to \code{\link[=uci_go]{uci_go()}}.}
}
\value{
A list containing the results of the analysis.
}
\description{
This function analyzes a chess position using a
\href{https://wbec-ridderkerk.nl/html/UCIProtocol.html}{UCI-compatible} chess
engine.
}
\details{
The function analyzes a chess position represented by a series of
chess moves in either short algebraic notation (SAN) or UCI long algebraic
notation (LAN), and uses a UCI-compatible chess engine to analyze the
position. The results include the best move, score, and best lines.
}
\examples{
# To run the example code, place a UCI-compatible chess engine in the
# `bigchess` subdirectory `/inst/extdata/engine`, or replace the line below
# with a valid path:
# engine_path <- "/put/your/own/engine/path/here"
engine_path <- find_engine()
if (!is.null(engine_path)) {
  analyze_position(engine_path, san = "1. e4", depth = 2)
} else {
 stop(paste0('To run the examples, install a chess engine in /inst/extdata/',
             'engine,\n or replace engine_path with the path to an engine.'))
}

}
\seealso{
\code{\link[=analyze_game]{analyze_game()}}
}
