% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.moves.R
\name{checking.moves}
\alias{checking.moves}
\title{Determine possible moves for a chess piece}
\usage{
checking.moves(position, r1, c1)
}
\arguments{
\item{position}{A matrix representing the current state of the chess board.}

\item{r1}{The row index of the piece for which to check possible moves.}

\item{c1}{The column index of the piece for which to check possible moves.}
}
\value{
A matrix where each row represents a possible destination square for
the piece. The first column stores the row indices and the second column
stores the column indices of these squares.
}
\description{
This function calculates the potential moves for a specific chess piece based
on its current position on the board.
}
\details{
The function takes into account the type of piece (e.g., rook,
queen, bishop, knight, pawn, king) and its current position to determine
all possible destination squares.
}
\examples{
# Determine possible moves for the white knight on b1 at the start of a game.
position <- position.start()
checking.moves(position = position, r1 = 1, c1 = 2)
}
