% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambiguating.move.R
\name{disambiguating.move}
\alias{disambiguating.move}
\title{Disambiguate chess moves}
\usage{
disambiguating.move(position, r1, c1, r2, c2)
}
\arguments{
\item{position}{A matrix representing the current position on the chess
board.}

\item{r1}{An integer representing the row of the piece before the move.}

\item{c1}{An integer representing the column of the piece before the move.}

\item{r2}{An integer representing the row of the piece after the move.}

\item{c2}{An integer representing the column of the piece after the move.}
}
\value{
A string that disambiguates the move. If no disambiguation is needed,
it returns an empty string.
}
\description{
When two or more identical chess pieces are able to move to the same square,
chess notation must somehow distinguish which piece was moved. In standard
algebraic notation (SAN), the ambiguity is resolved by adding row or column
names as needed to specify which piece moved. This function generates the
string of row or column names needed to specify the move in SAN.
}
\details{
The function evaluates the current state of the game and the
positions of the pieces before and after the move. It then generates a
string that uniquely identifies which piece made the move.
}
\examples{
# Initialize a chess board
position <- position.start()
# Add a white knight at e3
position[5, 5] <- 3
# Disambiguate a knight move from b1 (8, 2) to c3 (6, 3)
disambiguating.move(position, 8, 2, 6, 3)
}
