% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.R
\docType{data}
\name{eco}
\alias{eco}
\title{ECO}
\format{A data frame with ECO openings}
\usage{
data(eco)
}
\description{
A dataset containing 2014 ECO (Encyclopedia of Chess Openings) openings
\itemize{
  \item ECO
  \item Opening
  \item Variation
  \item Movetext: Standard algebraic notation
  \item NMoves
  \item LAN: Movetext converted into long algebraic notation
}
}
\keyword{datasets}
