% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_moves.R
\name{extract_moves}
\alias{extract_moves}
\title{Extract the first N moves}
\usage{
extract_moves(movetext, N = 10, last.move = T)
}
\arguments{
\item{movetext}{A character vector of movetext.}

\item{N}{(default = 10) An integer that determines how many moves will be
extracted. Must be greater than 0.}

\item{last.move}{(default TRUE) A Boolean indicating whether to calculate the
last move.}
}
\value{
A data frame containing the first N moves for white and for black.
}
\description{
This function takes a PGN movetext string (or string vector) and extracts the
first N moves into a data frame.
}
\details{
The function requires the standard English piece notations: pawn =
"P" (often not used), knight = "N", bishop = "B", rook = "R", queen = "Q",
and king = "K". It returns a data frame containing the first N moves for
white and for black, named as W1, B1, W2 and so on, up to WN and BN (where
N is the input argument). If N is greater than the total moves number then
NA's are generated.
}
\examples{
extract_moves("1. e4 e5 2. Nf3 Nf5 3. d5 ",N = 3)
# e4 e5 Nf3 Nf5 d5 NA TRUE
extract_moves("1. e4 e5 2. Nf3 Nf5 3. d5 ",N = 3, last.move = TRUE)
# e4 e5 Nf3 Nf5 d5 NA d5 TRUE

}
