% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_engine.R
\name{find_engine}
\alias{find_engine}
\title{Find a chess engine}
\usage{
find_engine()
}
\value{
A string containing the path to an engine, or \code{NULL} if no engine is
found.
}
\description{
A helper function allowing unit tests to run using a user-supplied chess
engine that is compatible with the local hardware/OS.
}
\details{
The \code{find_engine()} function returns the path to a chess engine that
has been placed in the \code{bigchess} subdirectory \verb{/inst/extdata/engine}. The
purpose is to allow unit tests and examples to use a chess engine without
having to edit every test and example to add the engine's path by hand.

Instead, a UCI-compatible chess engine suitable for the user's machine may
be placed in the \verb{/inst/extdata/engine} subdirectory, and the
\code{find_engine()} function can be used wherever an example or unit test would
otherwise require editing to add the path to an engine.

When an engine is found the function will return a string with the engine's
full path. If more than one engine is found, only the first one will be
used. If no engine is found, \code{find_engine()} will return \code{NULL}.
}
\note{
The \verb{/engine} directory contains a text file named
'Engine_goes_here-README' explaining the purpose of the directory. If any
other files are placed in \verb{/inst/extdata/engine} that are not
UCI-compatible chess engines capable of running on the user's machine, any
code making use of \code{find_engine()} is expected to fail.
}
\examples{
find_engine()
}
