% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstTwoMoves.R
\docType{data}
\name{FirstTwoMoves}
\alias{FirstTwoMoves}
\title{Example dataset}
\format{A data frame with popular positions in classic chess}
\usage{
data(FirstTwoMoves)
}
\description{
A dataset containing 10,894 results after first two moves in 2,395,869 high-quality chess games played over the board by players with ELO > 2000.
Source data OTB-HQ.7z downloaded from:
https://sourceforge.net/projects/codekiddy-chess/
and converted to PGN in SCID software.
\itemize{
  \item Result:
  \item W1: White first move
  \item B1: Black first move
  \item W2: White second move
  \item B2: Black second move
  \item Freq: Number of games played in database
}
}
\keyword{datasets}
