% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move2lan.R
\name{move2lan}
\alias{move2lan}
\title{Convert a chess move to Long Algebraic Notation (LAN)}
\usage{
move2lan(r1, c1, r2, c2, pr)
}
\arguments{
\item{r1}{An integer representing the row of the piece before the move.}

\item{c1}{An integer representing the column of the piece before the move.}

\item{r2}{An integer representing the row of the piece after the move.}

\item{c2}{An integer representing the column of the piece after the move.}

\item{pr}{An optional parameter representing a promotion piece.}
}
\value{
A string representing the move in Long Algebraic Notation (LAN).
}
\description{
This function converts a chess move from coordinate format to Long Algebraic
Notation (LAN).
}
\details{
The function takes the row and column of the piece before and after
the move, as well as an optional parameter for pawn promotion. If a pawn is
promoted during the move, the promotion piece should be specified as \code{2}
(bishop), \code{3} (knight), \code{4} (rook), or \code{5} (queen). If there is no
promotion, set \code{pr} to \code{NA}.
}
\examples{
move2lan(2, 5, 4, 5, NA) # Returns "e2e4"
}
