% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move2san.R
\name{move2san}
\alias{move2san}
\title{Convert a chess move to Standard Algebraic Notation (SAN)}
\usage{
move2san(position, r1, c1, r2, c2, p)
}
\arguments{
\item{position}{A matrix representing the current position on the chess
board.}

\item{r1}{An integer representing the row of the piece before the move.}

\item{c1}{An integer representing the column of the piece before the move.}

\item{r2}{An integer representing the row of the piece after the move.}

\item{c2}{An integer representing the column of the piece after the move.}

\item{p}{An optional parameter representing a promotion piece. If a pawn is
promoted as part of the move, \code{p} should be one of \code{2} (bishop), \code{3}
(knight),  \code{4} (rook), or \code{5} (queen). If there is no promotion, \code{p} should
be \code{NA}.}
}
\value{
A string representing the move in Standard Algebraic Notation (SAN).
}
\description{
This function takes a chess position and move coordinates, and returns the
move in Standard Algebraic Notation (SAN).
}
\details{
The function evaluates the current state of the game, the positions
of the pieces before and after the move, and any potential promotion. It
then generates a string that uniquely identifies the move in SAN.
}
\examples{
# Initialize a chess board
position <- position.start()
# Convert a move from (5, 7) to (5, 5) to SAN
move2san(position, 5, 7, 5, 5, NA) # Returns "e4"
}
