% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_profile.R
\name{player_profile}
\alias{player_profile}
\title{Compute a player's move frequency profile}
\usage{
player_profile(df, player)
}
\arguments{
\item{df}{A data frame of chess games.}

\item{player}{A string specifying the name of the player.}
}
\value{
A data frame where each row represents a game, and move counts for
each type of piece are arranged by column.
}
\description{
This function shows how often a player moved each type of chess piece.
}
\details{
The function accepts a data frame of chess games typically obtained
using \code{read.pgn()} or \code{read.pgn.ff()}. The function returns a data frame
where each row corresponds to a game, and move counts for each type of
piece are arranged by column.
}
\examples{
# Generate a profile for Kasparov
f <- system.file("extdata", "Kasparov.gz", package = "bigchess")
con <- gzfile(f, encoding = "latin1")
df <- read.pgn(con, quiet = TRUE, ignore.other.games = TRUE)
df_pp <- player_profile(df, "Kasparov")

# Generate a profile for Carlsen
f <- system.file("extdata", "Carlsen.gz", package = "bigchess")
con <- gzfile(f, encoding = "latin1")
df <- read.pgn(con, quiet = TRUE, ignore.other.games = TRUE)
df_pp <- player_profile(df, "Carlsen")

}
