% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_eco.R
\name{plot_tree_eco}
\alias{plot_tree_eco}
\title{Plot the tree for an ECO tree table}
\usage{
plot_tree_eco(tr, main = "", add.lines = TRUE, add.labels = TRUE)
}
\arguments{
\item{tr}{A data frame containing ECO tree data.}

\item{main}{(default = "") A string for the main title.}

\item{add.lines}{(default = TRUE) A Boolean indicating whether to add
weighted mean lines.}

\item{add.labels}{(default = TRUE) A Boolean indicating whether to add
labels.}
}
\value{
A barplot of the percentage of wins, losses, and draws.
}
\description{
This function creates a bar plot of the outcomes of a set of chess games.
}
\details{
The function takes a data frame of ECO tree data from \code{tree_eco()}
as an argument. The bar plot shows the percentage of games that resulted in
a win, loss, or draw. The weighted mean of the percentages is also shown.
}
\examples{
f <- system.file("extdata", "Kasparov.gz", package = "bigchess")
con <- gzfile(f, encoding = "latin1")
df <- read.pgn(con, quiet = TRUE, stat.moves = FALSE, add.tags = "ECO")
tr <- tree_eco(subset(df, W1 == "e4"), 20)
plot_tree_eco(tr, "1. e4 ... ?")

}
