% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_move.R
\name{plot_tree_move}
\alias{plot_tree_move}
\title{Plot the tree for a move tree table}
\usage{
plot_tree_move(tr, main = "", add.lines = TRUE, add.labels = TRUE)
}
\arguments{
\item{tr}{A data frame containing move tree data.}

\item{main}{(default = "") A string for the main title.}

\item{add.lines}{(default = TRUE) A Boolean indicating whether to add
weighted mean lines.}

\item{add.labels}{(default = TRUE) A Boolean indicating whether to add
labels.}
}
\value{
A barplot of the percentage of wins, losses, and draws based on
moves.
}
\description{
This function creates a bar plot of the outcomes of a set of chess games
based on a move tree.
}
\details{
The function takes a data frame of move tree data from \code{tree_move()}
as an argument. The bar plot shows the percentage of games that resulted in
a win, loss, or draw based on the moves made. The weighted mean of the
percentages is also shown.
}
\examples{
f <- system.file("extdata", "Kasparov.gz", package = "bigchess")
con <- gzfile(f, encoding = "latin1")
df <- read.pgn(con, quiet = TRUE, stat.moves = FALSE)
tr <- tree_move(subset(df,W1=="e4"),"B1")
plot_tree_move(tr,"1. e4 ... ?")

# Plot tree move openings in aggregated data
tr <- tree_move(FirstTwoMoves,"W1")
plot_tree_move(tr,paste0("1. ... ?\n",sum(FirstTwoMoves$Freq)," total games"))

}
