% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.move.R
\name{position.move}
\alias{position.move}
\title{Move a chess piece}
\usage{
position.move(position, r1, c1, r2, c2, p = NA)
}
\arguments{
\item{position}{A matrix representing the current state of the chess board.}

\item{r1, c1}{The rank and file of the piece's current position.}

\item{r2, c2}{The rank and file of the piece's destination.}

\item{p}{(optional) The piece to be promoted to, represented as a number. If
not provided, no promotion occurs.}
}
\value{
A matrix representing the updated state of the chess board after
executing the move.
}
\description{
This function moves a chess piece from its current position to a specified
destination on the chess board.
}
\details{
The function first identifies the pieces at the source and
destination positions. It then moves the piece from the source to the
destination.  It also handles special moves such as castling, en passant,
and promotion. If the move is a castling move (identified by a king moving
more than one square), it adjusts the rook's position based on whether it's
short or long castling. If the move is an en passant (identified by a pawn
moving diagonally to an empty square), it removes the opponent's pawn. If a
promotion is specified, it changes the pawn at the destination to the
specified piece.
}
\examples{
# Initialize a chess board
position <- position.start()
# Execute the move e4
updated_position <- position.move(position, r1=7, c1=5, r2=5, c2=5)

}
