% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible.moves.R
\name{possible.moves}
\alias{possible.moves}
\title{Compute possible moves for a piece}
\usage{
possible.moves(position, r1, c1)
}
\arguments{
\item{position}{A matrix representing the current state of the chess board.}

\item{r1, c1}{The rank and file of the piece's current position.}
}
\value{
A matrix where each row represents a possible move. Each row contains
two elements: the rank and file of the destination square.
}
\description{
This function calculates all possible moves for a specific piece on the chess
board. It takes into account the type of the piece and the current state of
the board.
}
\details{
The function evaluates the current state of the game and the
position of the piece in question. It then generates all potential moves
for this piece, considering the rules of movement for that particular
piece.
}
\examples{
# Initialize a chess board
position <- position.start()
# Compute possible moves for a piece at (2, 4)
possible.moves(position, r1 = 2, c1 = 4)
}
