% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pos.R
\name{print.pos}
\alias{print.pos}
\title{Print a chessboard}
\usage{
\method{print}{pos}(position)
}
\arguments{
\item{position}{An 8 x 8 numeric matrix of a chessboard.}
}
\value{
\code{NULL}
}
\description{
This function prints a chessboard to the console.
}
\details{
The function accepts an 8 x 8 matrix representing a chess position
as its argument. Each cell corresponds to a square, with numeric codes for
each piece, and with the sign indicating color (- for black): 0 = empty, 1
= pawn, 2 = bishop, 3 = knight, 4 = rook, 5 = queen, 6 = king. The pieces
are printed using Unicode.
}
\examples{
# Initialize a chessboard
position <- position.start()
# Print the chessboard
print.pos(position)

}
