% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.pgn.dt.R
\name{read.pgn.dt}
\alias{read.pgn.dt}
\title{Reads a PGN file into a data.table}
\usage{
read.pgn.dt(con, batch.size = 10^6, ignore.other.games = F, ...)
}
\arguments{
\item{con}{A string with a path to a file or URL, or a connection object.}

\item{batch.size}{(default = 10^6) An integer of the number of lines to read
per batch.}

\item{ignore.other.games}{(default = FALSE) A Boolean indicating if games
where the result is "*", (i.e. ongoing games) should be excluded.}

\item{...}{Further arguments passed to \code{read.pgn()} (excluding
\code{ignore.other.games} and \code{big.mode}).}
}
\value{
A message stating the number of games imported.
}
\description{
This function reads a PGN file into a data.table to more efficiently
manage and analyze a large collection of chess games.
}
\details{
The \code{con} argument can be a string with the path to a file to read,
a connection object, or a URL. The function reads the PGN files in batches
and writes them to a data.table. The size of the batches can be
specified by the user.
}
\examples{
\dontrun{
# Load required package
library(data.table)

# Specify file and connection
f <- system.file("extdata", "Carlsen.gz", package = "bigchess")
con <- gzfile(f, "rbt", encoding = "latin1")

# Read PGN data into data.table
dt <- read.pgn.dt(con, stat.moves = FALSE)

# Query the data.table
nrow(dt) #2410

# Example with different types of connections (also gz or zip files)
# Connection argument is passed directly to readLines(con, batch.size)
# So if total number of lines to read is greater than batch.size,
# use the correct opening mode for readLines depending on platform:

# Windows ('rb' opening mode for loop over readLines):
con <- gzfile(system.file("extdata", "Carlsen.gz", package = "bigchess"),
  "rb", encoding = "latin1")

dt <- read.pgn.dt(con)
}
}
