% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/san.move2move.R
\name{san.move2move}
\alias{san.move2move}
\title{Convert a chess move from standard algebraic notation (SAN) to coordinates}
\usage{
san.move2move(position, san.move, p)
}
\arguments{
\item{position}{A matrix representing the current position on the chess
board.}

\item{san.move}{A string representing the move in standard algebraic notation
(SAN).}

\item{p}{An integer representing the player (1 for white, -1 for black).}
}
\value{
A named integer vector representing the move in coordinate format.
The vector contains four or five elements: the row and column of the piece
before the move, the row and column of the piece after the move, and an
optional promotion piece. If a pawn is promoted as part of the move, the
fifth element should be one of \code{2} (bishop), \code{3} (knight), \code{4} (rook), or
\code{5} (queen). If there is no promotion, the vector will have length 4. The
names of the elements represent their coordinates in SAN.
}
\description{
This function converts a chess move from SAN to a more computational-friendly
coordinate format.
}
\details{
The function takes a chess position and a move in SAN as input. The
row numbers in the output are inverted compared to the traditional chess
board, with 1 at the top and 8 at the bottom. The output is a named integer
vector of length 4 or 5, depending on whether a pawn promotion occurs
during the move.
}
\examples{
position <- position.start()
san.move2move(position, "e4", 1) # Returns c(7, 5, 5, 5) with names c("", "", "4", "e")
}
