% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/san2lan.R
\name{san2lan}
\alias{san2lan}
\title{Convert chess moves from SAN to LAN}
\usage{
san2lan(movetext.san)
}
\arguments{
\item{movetext.san}{A string of chess moves in SAN.}
}
\value{
A string of chess moves in LAN.
}
\description{
This function converts a string of chess moves in standard algebraic notation
(SAN) into UCI long algebraic notation (LAN), which is used by chess engines.
}
\details{
The \code{san2lan} function takes a string of chess moves in SAN and
converts it to LAN. The string should not include comments, variants, etc.
}
\note{
This function assumes that the moves are valid and does not check for
errors or illegal moves.
}
\examples{
san2lan("1. e4 e5 2. Nf3 Nf5 3. d5 ")

}
