% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_moves.R
\name{stat_moves}
\alias{stat_moves}
\title{Count the number of times each type of piece moved}
\usage{
stat_moves(movetext, sides = "both")
}
\arguments{
\item{movetext}{A character vector of movetext from a PGN.}

\item{sides}{(default = "both") A string indicating which side's moves to
count ("white", "black", or "both").}
}
\value{
A data frame containing move count statistics for white and black, as
well as the total.
}
\description{
The function creates a data frame containing the number of times each type of
piece was moved.
}
\details{
The function counts the number of moves for each type of chess piece
(pawn, knight, bishop, rook, queen, and king) for both white and black
players, and calculates the total number of moves for each side (white,
black), as well as the combined total. The results are returned in a data
frame. The \code{movetext} argument must use The standard English piece
abbreviations: pawn = "P" (optional; often omitted), knight = "N", bishop =
"B", rook = "R", queen = "Q", and king = "K".
}
\examples{
stat_moves("1. e4 e5 2. Nf3 Nf5 3. d5 ")

}
