% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_moves.R
\name{stat_moves_black}
\alias{stat_moves_black}
\title{Count the number of times each type of black piece was moved}
\usage{
stat_moves_black(moves)
}
\arguments{
\item{moves}{A string representing a sequence of chess moves.}
}
\value{
A data frame containing the counts of moves for each type of black
chess piece.
}
\description{
This function takes a sequence of chess moves and calculates the number of
moves for each type of black chess piece (Bishop, King, Knight, Queen, Rook).
The input moves should be a string of chess moves.
}
\examples{
moves <- "1. e4 e5 2. Nf3 Nc6 3. Bb5 a6 4. Ba4"
black_move_counts <- stat_moves_black(moves)

}
