% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_eco.R
\name{tree_eco}
\alias{tree_eco}
\title{Compute an ECO tree}
\usage{
tree_eco(df, topn = 0)
}
\arguments{
\item{df}{A data frame containing the 'ECO' and 'Result' columns.}

\item{topn}{(default = 0) An integer indicating how many openings to include.
Set to 0 to include all openings.}
}
\value{
A data frame containing 'White_score' (White's winning percentage),
'Draws_percent' (draw percentage), 'Black_score' (Black's winning
percentage), and 'N' (number of games). The data frame is sorted by the
power of ECO (White_score * N), descending.
}
\description{
This function creates a data frame of the outcomes of a set of chess games.
}
\details{
This function takes a data frame with 'ECO' and 'Result' columns,
where 'ECO' represents the opening code and 'Result' denotes the game
result ('1-0' for White win, '1/2-1/2' for draw, '0-1' for Black win). The
function aggregates the data to calculate the winning percentages and
frequencies for each opening code.
}
\examples{
f <- system.file("extdata", "Kasparov.gz", package = "bigchess")
con <- gzfile(f, encoding = "latin1")
df <- read.pgn(con, quiet = TRUE, stat.moves = FALSE, add.tags = "ECO")
# Get the top 10 openings, by winning percentages
result <- tree_eco(df, topn = 10)

}
