% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uci_debug.R
\name{uci_debug}
\alias{uci_debug}
\title{Sending command debug for chess engine}
\usage{
uci_debug(engine, on = TRUE)
}
\arguments{
\item{engine}{engine object}

\item{on}{boolean default TRUE}
}
\value{
engine object
}
\description{
Sending command debug for chess engine. Info about debug command from http://wbec-ridderkerk.nl/html/UCIProtocol.html
switch the debug mode of the engine on and off. In debug mode the engine should sent additional infos to the GUI, e.g. with the "info string" command, to help debugging, e.g. the commands that the engine has received etc. This mode should be switched off by default and this command can be sent any time, also when the engine is thinking.
}
