% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uci_ponderhit.R
\name{uci_ponderhit}
\alias{uci_ponderhit}
\title{Sending command ponderhit for chess engine}
\usage{
uci_ponderhit(engine)
}
\arguments{
\item{engine}{engine object}
}
\value{
engine object
}
\description{
Sending command ponderhit for chess engine. Info about ponderhit command from http://wbec-ridderkerk.nl/html/UCIProtocol.html
the user has played the expected move. This will be sent if the engine was told to ponder on the same move the user has played. The engine should continue searching but switch from pondering to normal search.
}
