% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uci_read.R
\name{uci_read}
\alias{uci_read}
\title{Read current stdout from chess engine}
\usage{
uci_read(engine)
}
\arguments{
\item{engine}{engine object}
}
\value{
engine object
}
\description{
Read current stdout from chess engine
}
\examples{
\donttest{
# Linux (make sure you have executable permission):
engine_path <- "./stockfish_10_x64"
# Windows
# engine_path <- "./stockfish_10_x64.exe"
e <- uci_engine(engine_path)
e <- uci_read(e)
e$temp
uci_quit(e)}
}
