% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uci_register.R
\name{uci_register}
\alias{uci_register}
\title{Sending command register for chess engine}
\usage{
uci_register(engine, later = TRUE, name = NULL, code = NULL)
}
\arguments{
\item{engine}{engine object}

\item{later}{boolean default TRUE}

\item{name}{string}

\item{code}{string}
}
\value{
engine object
}
\description{
Sending command register for chess engine. Info about register command from http://wbec-ridderkerk.nl/html/UCIProtocol.html
this is the command to try to register an engine or to tell the engine that registration will be done later. This command should always be sent if the engine	has send "registration error" at program startup.
}
