% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uci_uci.R
\name{uci_uci}
\alias{uci_uci}
\title{Sending command uci for chess engine}
\usage{
uci_uci(engine)
}
\arguments{
\item{engine}{engine object}
}
\value{
engine object
}
\description{
Sending command uci for chess engine. Info about uci command from http://wbec-ridderkerk.nl/html/UCIProtocol.html
tell engine to use the uci (universal chess interface), this will be send once as a first command after program boot to tell the engine to switch to uci mode. After receiving the uci command the engine must identify itself with the "id" command and sent the "option" commands to tell the GUI which engine settings the engine supports if any. After that the engine should sent "uciok" to acknowledge the uci mode. If no uciok is sent within a certain time period, the engine task will be killed by the GUI.
}
