% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lan2san.R
\name{lan2san}
\alias{lan2san}
\title{Movetext conversion from LAN to SAN}
\usage{
lan2san(movetext.lan)
}
\arguments{
\item{movetext.lan}{movetext string in long algebraic notation (LAN), but without comments, variants etc.}
}
\value{
movetext in standard algebraic notation
}
\description{
Convert LAN movetext (long algebraic notation used by chess engines) to SAN movetext (standard algebraic notation required by FIDE)
}
\examples{
lan2san("e2e4 c7c5")
}
