% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_moves.R
\name{n_moves}
\alias{n_moves}
\title{Compute number of moves}
\usage{
n_moves(movetext)
}
\arguments{
\item{movetext}{movetext string (or string vector)}
}
\value{
n integer (or integer vector)
}
\description{
Compute total number of moves given movetext string (or string vector)
}
\examples{
n_moves(c("1. e4 e5 2. Nf3 Nf5 3. d5 ","1. d4 d5"))
# 3 1
}
