% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/san2lan.R
\name{san2lan}
\alias{san2lan}
\title{Movetext conversion from SAN to LAN}
\usage{
san2lan(movetext.san)
}
\arguments{
\item{movetext.san}{movetext string in standard algebraic notation (SAN) required by FIDE, but without comments, variants etc.}
}
\value{
movetext in long algebraic notation
}
\description{
Convert SAN movetext (FIDE) to LAN movetext (used by chess engines)
}
\examples{
san2lan("1. e4 e5 2. Nf3 Nf5 3. d5 ")
}
