% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_moves.R
\name{stat_moves}
\alias{stat_moves}
\title{Extract statistics of moves}
\usage{
stat_moves(movetext, sides = "both")
}
\arguments{
\item{movetext}{movetext string (or string vector). The standard English values are required: pawn = "P" (often not used), knight = "N", bishop = "B", rook = "R", queen = "Q", and king = "K".}

\item{sides}{"both" (default),"white" or "black"}
}
\value{
Data frame containing moves count statistics for white and for black and total.
}
\description{
Extract statistics of moves (counts figure moves) from movetext vector into data frame
}
\examples{
stat_moves("1. e4 e5 2. Nf3 Nf5 3. d5 ")
}
